+++ ui/app/components/currency-input/tests/currency-input.container.test.js
@@ -52,12 +52,8 @@ describe('CurrencyInput container', () => {
       assert.deepEqual(mergeProps(mockStateProps, mockDispatchProps, {}), {
         conversionRate: 280.45,
         currentCurrency: 'usd',
-<<<<<<< HEAD
-        nativeCurrency: 'ETH',
-        suffix: 'ETH',
-=======
+        nativeCurrency: 'DEX',
         suffix: 'DEX',
->>>>>>> Update Dexon wording
       })
     })
   })